Move StudyPlanViewState to view package
