+++ shared/src/commonMain/kotlin/org/hyperskill/app/study_plan/injection/StudyPlanComponent.kt
@@ -1,7 +1,7 @@
 package org.hyperskill.app.study_plan.injection
 
 import org.hyperskill.app.study_plan.presentation.StudyPlanFeature
-import org.hyperskill.app.study_plan.presentation.StudyPlanViewState
+import org.hyperskill.app.study_plan.view.StudyPlanViewState
 import ru.nobird.app.presentation.redux.feature.Feature
 
 interface StudyPlanComponent {
