+++ shared/src/commonMain/kotlin/org/hyperskill/app/study_plan/injection/StudyPlanComponentImpl.kt
@@ -2,7 +2,7 @@ package org.hyperskill.app.study_plan.injection
 
 import org.hyperskill.app.core.injection.AppGraph
 import org.hyperskill.app.study_plan.presentation.StudyPlanFeature
-import org.hyperskill.app.study_plan.presentation.StudyPlanViewState
+import org.hyperskill.app.study_plan.view.StudyPlanViewState
 import ru.nobird.app.presentation.redux.feature.Feature
 
 class StudyPlanComponentImpl(private val appGraph: AppGraph) : StudyPlanComponent {
